requirements.txt
passlib~=1.7.4
python-jose~=3.5.0
python-dotenv~=1.1.1
fastapi~=0.116.1
app~=0.0.1
pydantic~=2.11.7
sqlalchemy~=2.0.42
fastapi
uvicorn
passlib[bcrypt]
python-jose
pydantic

requests~=2.32.4
db~=0.1.1
openai~=1.98.0
together~=1.5.21
pillow~=11.3.0
pdfplumber~=0.11.7
pytesseract~=0.3.13
openai-whisper~=20250625
PyPDF2~=3.0.1